requirements.txt
Flask
Flask-RESTful
Flask-Cors
flask-sqlalchemy
Flask-Session
Flask-Migrate
psycopg2
pyjwt
bcrypt